Confirm contact details - refactor Address LookUp
