+++ src/main/steps/applicant/confirm-contact-details/checkanswers/ConfirmContactDetailsGetController.ts
@@ -9,7 +9,6 @@ import {
   getConfidentialData,
   validateDataCompletion,
 } from './../../../../steps/common/confirm-contact-details/checkanswers/ConfirmContactDetailsGetController';
-
 @autobind
 export default class ConfirmContactDetailsGetController extends GetController {
   public async get(req: AppRequest, res: Response): Promise<void> {
