+++ src/main/steps/applicant/confirm-contact-details/checkanswers/controller/ApplicantConfirmContactDetailsGetController.ts
@@ -1,10 +1,10 @@
 import autobind from 'autobind-decorator';
 import { Response } from 'express';
 
-import { FieldPrefix } from '../../../../app/case/case';
-import { AppRequest } from '../../../../app/controller/AppRequest';
-import { CommonContent } from '../../../../steps/common/common.content';
-import { ConfirmContactDetailsGetController } from '../../../../steps/common/confirm-contact-details/checkanswers/ConfirmContactDetailsGetController';
+import { FieldPrefix } from '../../../../../app/case/case';
+import { AppRequest } from '../../../../../app/controller/AppRequest';
+import { CommonContent } from '../../../../../steps/common/common.content';
+import { ConfirmContactDetailsGetController } from '../../../../../steps/common/confirm-contact-details/checkanswers/ConfirmContactDetailsGetController';
 
 export type PageContent = Record<string, unknown>;
 export type TranslationFn = (content: CommonContent) => PageContent;
