+++ src/main/steps/applicant/confirm-contact-details/checkanswers/controller/ApplicantConfirmContactDetailsPostController.ts
@@ -1,10 +1,10 @@
 import autobind from 'autobind-decorator';
 import { Response } from 'express';
 
-import { FieldPrefix } from '../../../../app/case/case';
-import { AppRequest } from '../../../../app/controller/AppRequest';
-import { CommonContent } from '../../../../steps/common/common.content';
-import { ConfirmContactDetailsPostController } from '../../../../steps/common/confirm-contact-details/checkanswers/ConfirmContactDetailsPostController';
+import { FieldPrefix } from '../../../../../app/case/case';
+import { AppRequest } from '../../../../../app/controller/AppRequest';
+import { CommonContent } from '../../../../../steps/common/common.content';
+import { ConfirmContactDetailsPostController } from '../../../../../steps/common/confirm-contact-details/checkanswers/ConfirmContactDetailsPostController';
 
 export type PageContent = Record<string, unknown>;
 export type TranslationFn = (content: CommonContent) => PageContent;
