+++ src/main/steps/respondent/confirm-contact-details/checkanswers/controller/RespondentConfirmContactDetailsGetController.test.ts
@@ -1,10 +1,10 @@
 import { RespondentConfirmContactDetailsGetController } from './RespondentConfirmContactDetailsGetController';
 jest.mock('autobind-decorator');
 jest.mock('express');
-jest.mock('../../../../app/case/case');
-jest.mock('../../../../app/controller/AppRequest');
-jest.mock('../../../../steps/common/common.content');
-jest.mock('../../../../steps/common/confirm-contact-details/checkanswers/ConfirmContactDetailsGetController');
+jest.mock('../../../../../app/case/case');
+jest.mock('../../../../../app/controller/AppRequest');
+jest.mock('../../../../../steps/common/common.content');
+jest.mock('../../../../../steps/common/confirm-contact-details/checkanswers/ConfirmContactDetailsGetController');
 
 describe('RespondentConfirmContactDetailsGetController', () => {
   let controller;
